+++ CHANGELOG.md
@@ -7,6 +7,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org).
 
 
+[unreleased]
+------------
+
+### Added
+- Ability to specify a RefreshJob class name
+
+
 [v0.3.0]
 --------